requirements.txt
streamlit>=1.31.0
pillow>=10.0.0